Get linkage disequilibrium

Get LD from a group of snps or a single tag snp to all snps in a window.

Usage

get_ld_in_window(
  qtl.df = NULL,
  tag.snp = NULL,
  window,
  plink.path = NULL,
  geno.bed,
  in.dir,
  out.dir = NULL,
  pvals.in.log = T,
  verbose = TRUE
)

Arguments

qtl.df

data.frame, table that includes list of snps to calculate LD to with columns (CHR, POS, PVAL), corresponding to (chromosome, physical position, and p-value). QTL are typically defined as hits grouped by LD by something like plink --clump

tag.snp

character, marker.ID of snp around which to calculate LD. In the form 'CHR-POS'

window

numeric, a physical distance to determine which snps to calculate LD for. Snps within this distance to the tag.snp or qtl.df snps will be calculated. The entire size of the region centered on the tag snp within which LD will be calculated is 2*window.

plink.path

character, optional, path to plink2 executable. Will overide option set by set_plink_path.

geno.bed

character, prefix of genotype files in plink (bed/bim/fam) format. Do not include ".bed" extension.

in.dir

character, directory where genotype files are located

out.dir

character, where to output some temporary files.

pvals.in.log

boolean, are p-values in PVAL column of qtl.df represented as -log10(p-value)?

verbose

boolean, if TRUE, output some status reports

Value

Named list with 2 items

table: table with marker.IDs (CHR-POS) and maximum LD in R2 for each snp to the snps in the qtl.df or LD to tag.snp

key.snp: marker.ID corresponding to the middle of the window, either the max(LOGPVAL) of qtl.df or tag.snp. useful to retain for downstream functions.

qtl.snps: marker.ID corresponding to the markers in the qtl.df. useful to retain for downstream functions.

Details

When supplying a group a snps, the ld retained for a given snp is the maximum LD of the given snp to the snps in the group.

References

Chang CC, Chow CC, Tellier LCAM, Vattikuti S, Purcell SM, Lee JJ (2015) Second-generation PLINK: rising to the challenge of larger and richer datasets. GigaScience, 4.

Examples

# set up inputs for example
plink.path <- bigsnpr::download_plink2()
tag.snp <- "Chr_05-6857045"
geno.bed <- "Setaria_shattering_example_pruned.bed"
out.dir <- file.path(tempdir(), "panvar_ex")
dir.create(out.dir, showWarnings = FALSE)

# run example
out <- get_ld_in_window(
    tag.snp = tag.snp,
    window = 25,
    plink.path = plink.path,
    geno.bed = geno.bed,
    in.dir = system.file("extdata", package="panvaR"),
    out.dir = out.dir)

# list
names(out)
#> [1] "table"                  "key.snp"                "key.snp_geno.formatted"
#> [4] "qtl.snps"              
# table
head(out$table)
#>        marker.ID CHR     POS         R2
#> 1 Chr_05-6832059   5 6832059 0.00679048
#> 2 Chr_05-6832638   5 6832638 0.00679048
#> 3 Chr_05-6833053   5 6833053 0.01793830
#> 4 Chr_05-6833177   5 6833177 0.92282300
#> 5 Chr_05-6833193   5 6833193 0.00513623
#> 6 Chr_05-6833227   5 6833227 0.01793830

# clean up
unlink(out.dir, recursive = TRUE)