sets global option "panvar_prefix" for output directory of intermediate panvar files. If not supplied will be sent to output of 'tempdir()'.

sets global option "panvar_prefix" for output directory of intermediate panvar files. If not supplied will be sent to output of 'tempdir()'.

Usage

set_panvar_prefix(prefix)

Arguments

prefix

character, a prefix to designate files from individual runs of the program.

Examples

# set_panvar_prefix("Phenotype1")